+++ adagucserverEC/CMakeLists.txt
@@ -192,9 +192,10 @@ add_library(
     CServerParams.h
     CStyleConfiguration.cpp
     CStyleConfiguration.h
-    CUniqueRequests/CURResult.h
+    CUniqueRequests/CURTypes.h
     CUniqueRequests/requestUtils.h
     CUniqueRequests/requestUtils.cpp
+    CUniqueRequests/CURUniqueRequests.h
     CUniqueRequests/CURUniqueRequests.cpp
     CXMLGen.cpp
     CXMLGen.h
