--- adagucserverEC/CUniqueRequests/CURConstants.h
@@ -1,4 +1,0 @@
-#include <map>
-#define CCUniqueRequests_MAX_DIMS 255
-typedef std::map<int, std::string> map_type_dimvalindex;
-typedef std::map<int, std::string>::iterator it_type_dimvalindex;
